move WriteTar check to write_hls and align Catapult behavior

+++ hls4ml/writer/vivado_accelerator_writer.py
@@ -405,10 +405,12 @@ class VivadoAcceleratorWriter(VivadoWriter):
         )
 
     def write_new_tar(self, model):
-        tarfile = model.config.get_output_dir() + '.tar.gz'
-        if os.path.exists(tarfile):
-            os.remove(tarfile)
-        super().write_tar(model)
+        write_tar = model.config.get_writer_config().get('WriteTar', False)
+        if write_tar:
+            tarfile = model.config.get_output_dir() + '.tar.gz'
+            if os.path.exists(tarfile):
+                os.remove(tarfile)
+            super().write_tar(model)
 
     def write_hls(self, model):
         """
